\name{qualV-package}
\alias{qualV-package}
\alias{qualV}

\title{Qualitative Validation Methods}

\description{Qualitative methods for model validation.}

\details{This package contains functions for a qualitative model comparison.
  Common quantitative deviance measures underestimate the
  similarity of patterns if there are shifts in time between measurement
  and simulation. Qualitative validation methods are additional methods
  to validate models, especially useful to compare the patterns of observed and
  simulated values.
  \cr\cr
  For a complete list of functions with individual help pages, use
  \code{library(help="qualV")}.
}

\references{
  Jachner, S., van den Boogaart, K.G. and Petzoldt, T. (2007) Statistical
  methods for the qualitative assessment of dynamic models with time
  delay (R package qualV), Journal of Statistical Software, 22(8), 1--30.
  \url{http://dx.doi.org/10.18637/jss.v022.i08}.
}
\keyword{misc}

